feat: handle language already exists error
