+++ packages/api-types/src/index.ts
@@ -1,43 +1,9 @@
-export interface LanguageResourceIdentifier {
-  type: 'language';
-  id: string;
+export * from './language';
+
+export interface Error {
+  code: string;
 }
 
-export interface Language extends LanguageResourceIdentifier {
-  attributes: {
-    name: string;
-  };
-}
-
-export interface LanguageWithLinks extends Language {
-  links: {
-    self: string;
-  };
-}
-
-export interface GetLanguagesResponseBody {
-  data: LanguageWithLinks[];
-  links: {
-    self: string;
-  };
-}
-
-export interface GetLanguageResponseBody {
-  data: LanguageWithLinks;
-}
-
-export interface PostLanguageRequestBody {
-  data: Language;
-}
-
-export interface PostLanguageResponseBody {
-  data: LanguageWithLinks;
-}
-
-export interface PatchLanguageRequestBody {
-  data: Language;
-}
-
-export interface PatchLanguageResponseBody {
-  data: LanguageWithLinks;
+export interface ErrorResponse {
+  errors: Error[];
 }
