+++ packages/api/helpers.ts
@@ -1,9 +1,8 @@
+import { ErrorResponse } from '@translation/api-types';
 import { NextApiRequest, NextApiResponse } from 'next';
 
 export type ApiRequest<Params = void> = Params extends void
   ? NextApiRequest
   : NextApiRequest & { query: Params };
 
-export type ApiResponse<Body = void> = NextApiResponse<
-  Body | { errors: { code: string }[] }
->;
+export type ApiResponse<Body = void> = NextApiResponse<Body | ErrorResponse>;
